requirements.txt
deepface==0.0.68
pinecone-client
opencv-python==4.5.5.64
Flask
numpy
Werkzeug
streamlit
Pillow
mtcnn
waitress
fastapi
uvicorn
python-multipart
gunicorn
# -e .